Moving from DMS to Noctalia Shell v5

+++ build_files/build.sh
@@ -4,7 +4,7 @@ set -ouex pipefail
 dnf remove -y vim-minimal vim-enhanced firefox firefox-langpacks gnome-keyring nano \
 nano-default-editor
 
-dnf install -y greetd greetd-selinux  \
+dnf install -y greetd greetd-selinux \
 kernel-modules-extra tcpdump symlinks appstream cups zip unzip wget2 bluez-cups dnf5-plugins \
 alsa-firmware alsa-tools-firmware alsa-utils pipewire pipewire-gstreamer flatpak flatpak-selinux \
 pipewire-alsa pipewire-pulseaudio pipewire-utils pipewire-jack-audio-connection-kit dhcp-client \
@@ -13,21 +13,25 @@ NetworkManager-openvpn-gnome openvpn systemd-container systemd-networkd fish Ima
 sane-backends-drivers-cameras sane-backends-drivers-scanners rootfiles dhcp-client \
 system-config-printer-udev system-config-printer-libs iwlwifi-mld-firmware iwlwifi-mvm-firmware
 
-dnf -y copr enable avengemedia/dms
-dnf -y copr enable avengemedia/danklinux
-dnf -y copr enable yalter/niri
-dnf config-manager addrepo --from-repofile=https://negativo17.org/repos/fedora-multimedia.repo
-dnf install  --setopt=install_weak_deps=False -y quickshell matugen cliphist network-manager-applet \
-dms dms-greeter material-symbols-fonts niri xwayland-satellite xdg-desktop-portal-gtk rar libdvdcss \
-grim slurp mako alacritty wl-clipboard wlsunset mate-polkit qt6-qtwayland adw-gtk3-theme x264 x265 \
+ln -sf /usr/bin/nvim /usr/bin/vi
+ln -sf /usr/bin/nvim /usr/bin/vim
+
+dnf install -y --nogpgcheck --repofrompath "terra,https://repos.fyralabs.com/terra/\$releasever" terra-release
+dnf config-manager --set-enabled terra-multimedia
+
+dnf install --setopt=install_weak_deps=False -y noctalia-git noctalia-greeter \
+material-symbols-fonts niri xwayland-satellite xdg-desktop-portal-gtk matugen rar libdvdcss cliphist \
+grim slurp mako ghostty wl-clipboard wlsunset mate-polkit qt6-qtwayland adw-gtk3-theme x264 x265 \
 ffmpeg gstreamer1-plugins-good gstreamer1-plugins-bad-free imv caja atril mousepad engrampa paperwork
 
-systemctl enable --now greetd
-systemctl enable --now firewalld
+systemctl enable greetd
+systemctl enable firewalld
 
-mv /usr/share/doc/niri/wiki/Layer‐Shell-Components.md /usr/share/doc/niri/wiki/Layer-Shell-Components.md
+if [ -f /usr/share/doc/niri/wiki/Layer‐Shell-Components.md ]; then
+    mv /usr/share/doc/niri/wiki/Layer‐Shell-Components.md /usr/share/doc/niri/wiki/Layer-Shell-Components.md
+fi
 
 dnf clean all
-rm -rf /var/tmp/dnf/cache/
+rm -rf /var/cache/dnf/ /var/tmp/dnf/
 ostree container commit
 rm -rf /tmp/*
